Enable check of working copy against arbitrary git object
This is not useful when developing on the latest version of the
`tisch-merge` branch but can be useful when trying to identify the
commit of `master` corresponding to the commit of `tisch-merge`
currently checked out.

E.g. checking out f801419 (that is on `tisch-merge` branch) then
running `./check.sh 6.06` (tag `6.06` is taken from `master` branch)
shows that such commit (from `tisch-merge`) is in line with the tag
(from `master`).