1 Cor 14:37: Fix accent
This version has ἐστὶν, but the version in
https://github.com/LogosBible/SBLGNT/data/sblgnt/text has ἐστίν.

Since SBLGNT/data/sblgntapp/text/1Cor.txt has the following entry:

1 Corinthians 14:37
37 ἐστίν Holmes ] ἐστὶν ἐντολή WH Treg NA28; εἰσὶν ἐντολαί RP

In the SBLGNT, ἐστίν is followed by a period.  According to the
Ancient Greek accent article on Wikipedia [1], the section on Oxytone
words / Change to grave, this means it should remaine acute; the
alternate notes match this pattern.

So overall it seems like ἐστίν is probably correct.

[1] https://en.wikipedia.org/wiki/Ancient_Greek_accent

Signed-off-by: George Dunlap <dunlapg@umich.edu>