2 Peter 2:10: Fix accent
This version has Τολμηταὶ, but the version in
https://github.com/LogosBible/SBLGNT/data/sblgnt/text has Τολμηταί.

In the SBLGNT, Τολμηταί is followed by a comma.  According to the
Ancient Greek accent article on Wikipedia [1], the section on Oxytone
words / Change to grave, this means it should remaine acute; the
alternate notes match this pattern.

The same verse at esv.org has a simlar pattern: Τολμηταί follwed by a
comma.

So overall it seems like ἐστίν is probably correct.

[1] https://en.wikipedia.org/wiki/Ancient_Greek_accent

Signed-off-by: George Dunlap <dunlapg@umich.edu>