Acts 2:29: Fix accent
This version has Δαυὶδ, but the version in
https://github.com/LogosBible/SBLGNT/data/sblgnt/text has Δαυίδ.

In the SBLGNT, Δαυίδ is followed by a comma.  According to the
Ancient Greek accent article on Wikipedia [1], the section on Oxytone
words / Change to grave, this means it should remaine acute.

The same text at [1] follows this pattern.

So it seems like Δαυίδ is probably correct.

[1] https://www.greekbible.com/

Signed-off-by: George Dunlap <dunlapg@umich.edu>