Hebrews 3:6: Fix accent
This version has ὅς, but the version in
https://github.com/LogosBible/SBLGNT/data/sblgnt/text has ὃς.

Then sblgntapp entry[1] regarding the verse says:

Hebrews 3:6
6 ὃς Holmes ] οὗ WH Treg NA28 RP
• ἐὰν WH Treg ] ἐάνπερ NA28 RP
• ἐλπίδος NA28 ] + μέχρι τέλους βεβαίαν WH Treg RP

In the SBLGNT, ὃς is followed by a non-encletic word.  According to the
Ancient Greek accent article on Wikipedia [2], the section on Oxytone
words / Change to grave, this means it should change to grave.

The texts at esv.org and greekbible.com both use οὗ, and so don't
provide a point of comparison.

Nonetheless, overall it seems like ὃς is probably correct.

[1] https://github.com/LogosBible/SBLGNT/blob/master/data/sblgntapp/text/Heb.txt
[2] https://en.wikipedia.org/wiki/Ancient_Greek_accent

Signed-off-by: George Dunlap <dunlapg@umich.edu>