John 7:34, 36: Fix accents
This version has εὑρήσετέ, but the version in
https://github.com/LogosBible/SBLGNT/data/sblgnt/text has εὑρήσετε.

Then sblgntapp entry[1] regarding the verses say:

John 7:34
34 εὑρήσετε Treg RP ] + με WH NA28

John 7:36
36 ὁ λόγος οὗτος WH Treg NA28 ] οὗτος ὁ λόγος RP
• εὑρήσετε Treg RP ] + με WH NA28

esv.org also has no accent at the end (in both cases), although
greekbible.com does (but has [με] in brackets afterwards).  On the
whole it seems likely that the lack of accent in sblgnt was
intentional.

[1] https://github.com/LogosBible/SBLGNT/blob/master/data/sblgntapp/text/John.txt

Signed-off-by: George Dunlap <dunlapg@umich.edu>