Mark 7:27: Fix accent
This version has γάρ, but the version in
https://github.com/LogosBible/SBLGNT/data/sblgnt/text has γὰρ.

This one took me down a long rabbit trail, since the texts at both
esv.org [1] and greekbible.com [2] have γάρ.  However, according to
the Wikipedia entry on Ancient Greek diacritics [3]:

<quote>

The grave is found, as an alternative to an acute, only on the last
syllable of a word. When a word such as ἀγαθός agathós 'good' with
final accent is followed by a pause (that is, whenever it comes at the
end of a clause, sentence, or line of verse),[1] or by an enclitic
word such as the weak form of ἐστίν estín 'is' (see below), the accent
is written as an acute:

 * ἀνὴρ ἀγαθός anḕr agathós '[a] good man'
 * ἀνὴρ ἀγαθός ἐστιν anḕr agathós estin '[he] is [a] good man'

However, when the word does not come before a pause or an enclitic,
the acute accent is replaced by a grave:

 * ἀγαθὸς ἄνθρωπος agathòs ánthrōpos 'a good person'

</quote>

In both esv.org and greekbible.com γάρ is indeed followed by ἐστίν;
but in SBLGNT, γὰρ is followed by καλόν.

All in all, it seems likely that γὰρ is correct.

[1] https://www.esv.org/Mark+7/
[2] https://www.greekbible.com/
[3] https://en.wikipedia.org/wiki/Ancient_Greek_accent

Signed-off-by: George Dunlap <dunlapg@umich.edu>