Matthew 15:15: Fix accent
This version has παραβολήν, but the version in
https://github.com/LogosBible/SBLGNT/data/sblgnt/text has παραβολὴν.

Considering again the rules regarding Oxytone words: Change to grave
from the Ancient Greek accent Wikipedia article [1], it would seem the
SBLGNT text would be correct.

The versions at the esv.org and greekbible.com seem to follow the
rule, using παραβολήν but at the end of the sentence.

All in all, it seems likely that παραβολὴν is correct.

[1] https://en.wikipedia.org/wiki/Ancient_Greek_accent

Signed-off-by: George Dunlap <dunlapg@umich.edu>