Matthew 26:36: Fix accent
This version has Γεθσημανὶ, but the version in
https://github.com/LogosBible/SBLGNT/data/sblgnt/text has Γεθσημανί.

Considering the rules regarding Oxytone words: Change to grave
from the Ancient Greek accent Wikipedia article [1], it would seem the
SBLGNT text would be correct, since it's followed by a comma.

The version at greekbible.com also has a comma, and also uses the
acute; the version at esv.org uses a grave, but isn't followed by a
comma or an enclined word

All in all, it seems likely that Γεθσημανί is correct.

[1] https://en.wikipedia.org/wiki/Ancient_Greek_accent

Signed-off-by: George Dunlap <dunlapg@umich.edu>